Fix menus
- Remove dropdown menu
- Show PresetCount & only preset in Accordion menu
- Show display name from presets
