+++ pkg/menu/gallery_menu.go
@@ -17,6 +17,7 @@ limitations under the License.
 package menu
 
 import (
+	"context"
 	"fmt"
 	gourl "net/url"
 	"path"
@@ -24,8 +25,10 @@ import (
 
 	"github.com/pkg/errors"
 	"gomodules.xyz/pointer"
+	core "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	rsapi "kmodules.xyz/resource-metadata/apis/meta/v1alpha1"
+	sharedapi "kmodules.xyz/resource-metadata/apis/shared"
 	"kmodules.xyz/resource-metadata/hub/resourceeditors"
 	"kubepack.dev/kubepack/pkg/lib"
 	"kubepack.dev/lib-helm/pkg/values"
@@ -123,9 +126,7 @@ func RenderGalleryMenu(kc client.Client, in *rsapi.Menu, opts *rsapi.RenderMenuR
 
 					if len(ed.Spec.Variants) == 1 {
 						// cp := mi
-						if name != resourceeditors.DefaultEditorName(mi.Resource.GroupVersionResource()) {
-							mi.Name = name
-						}
+						mi.Name = name
 						mi.Path = u.String()
 						mi.Preset = &ref.TypedLocalObjectReference
 						if len(ref.Icons) > 0 {
@@ -156,4 +157,26 @@ func RenderGalleryMenu(kc client.Client, in *rsapi.Menu, opts *rsapi.RenderMenuR
 	}
 
 	return &out, nil
+}
+
+func GetPresetName(
+	kc client.Client,
+	chartRef *sharedapi.ChartRepoRef,
+	vpsMap map[string]*chartsapi.VendorChartPreset,
+	ref core.TypedLocalObjectReference,
+) (string, error) {
+	if ref.Kind == chartsapi.ResourceKindVendorChartPreset {
+		ps, ok := vpsMap[ref.Name]
+		if !ok {
+			return "", fmt.Errorf("%s %s not found in chart %+v", chartsapi.ResourceKindVendorChartPreset, ref.Name, chartRef)
+		}
+		return ps.Name, nil
+	}
+
+	var ps chartsapi.ClusterChartPreset
+	err := kc.Get(context.TODO(), client.ObjectKey{Name: ref.Name}, &ps)
+	if err != nil {
+		return "", errors.Wrapf(err, "%s %s not found", chartsapi.ResourceKindClusterChartPreset, ref.Name)
+	}
+	return ps.GetDisplayName(), nil
 }
